pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.0"
description = ""
authors = ["Pavel Kalenchuk <pavel.kalenchuk@gmail.com>"]
packages = [
    {include = "gendiff"},
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-black = "^0.2.4"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"